requirements.txt
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.20.0
nltk>=3.6
layoutparser>=0.3.0
pyxdameraulevenshtein>=1.7.0
python-doctr>=0.5.0
transformers>=4.15.0
torch>=1.10.0
detectron2>=0.6
google-api-python-client>=2.0.0
beautifulsoup4>=4.9.0
requests>=2.25.0
farm-haystack>=1.8.0